requirements.txt
streamlit~=1.39.0
langchain~=0.3.4
langchain_community
langchain_core
tiktoken
torch>=1.10.0
transformers>=4.20.0
groq>=0.1.0
ollama~=0.3.3
huggingface-hub>=0.0.19
langchain-community~=0.3.3
langchain-core~=0.3.12
requests~=2.32.3
pillow~=10.4.0
pandas~=2.2.3
nltk~=3.9.1
tqdm~=4.66.5
spacy==3.5.0
